Create initial Omics Integrator 1 run method
